Create netspeed.py
This script takes the input of a CSV file (included in repository), extracts the data, and prints it to the console. When the user continues the program by entering "y" when prompted to continue, each material's trial average times (ms) will be averaged and compared to the control environment's average times (ms). The total decrease and percent decrease will be calculated and printed to the console. This is currently being used to uncover which material is best to stop wireless signals.