Create netstrength.py
This program takes the input of a CSV file (netstrength_data.csv in repository) containing the network strength data that I have collected. It then prints out the data in a neat data table. It then finds the average signal drop from the control environment, and the max signal drop from the control environment, and prints the results to the console.